+++ src/main/java/dbdr/DbdrApplication.java
@@ -8,8 +8,8 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @EnableJpaAuditing
 public class DbdrApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(DbdrApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(DbdrApplication.class, args);
+    }
 
 }
