--- src/main/java/dbdr/chart/domain/PhysicalWalk.java
@@ -1,21 +1,0 @@
-package dbdr.chart.domain;
-
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.Comment;
-
-@Embeddable
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class PhysicalWalk {
-    @Comment("산책 여부")
-    private boolean hasWalked;
-
-    @Comment("외출 동행")
-    private boolean hasCompanion;
-}
