--- src/main/java/dbdr/chart/dto/request/ChartDetailRequest.java
@@ -1,9 +1,0 @@
-package dbdr.chart.dto.request;
-
-public record ChartDetailRequest(
-        BodyManagementRequest bodyManagement,
-        NursingManagementRequest nursingManagement,
-        CognitiveManagementRequest cognitiveManagement,
-        RecoveryTrainingRequest recoveryTraining
-) {
-}
