--- src/main/java/dbdr/chart/dto/response/ChartDetailResponse.java
@@ -1,10 +1,0 @@
-package dbdr.chart.dto.response;
-
-public record ChartDetailResponse(
-        Long chartId,
-        BodyManagementResponse bodyManagement,
-        NursingManagementResponse nursingManagement,
-        CognitiveManagementResponse cognitiveManagement,
-        RecoveryTrainingResponse recoveryTraining
-) {
-}
