+++ src/main/java/dbdr/domain/chart/dto/request/ChartDetailRequest.java
@@ -1,0 +1,9 @@
+package dbdr.domain.chart.dto.request;
+
+public record ChartDetailRequest(
+        BodyManagementRequest bodyManagement,
+        NursingManagementRequest nursingManagement,
+        CognitiveManagementRequest cognitiveManagement,
+        RecoveryTrainingRequest recoveryTraining
+) {
+}
