+++ src/main/java/dbdr/domain/chart/dto/response/ChartDetailResponse.java
@@ -1,0 +1,10 @@
+package dbdr.domain.chart.dto.response;
+
+public record ChartDetailResponse(
+        Long chartId,
+        BodyManagementResponse bodyManagement,
+        NursingManagementResponse nursingManagement,
+        CognitiveManagementResponse cognitiveManagement,
+        RecoveryTrainingResponse recoveryTraining
+) {
+}
