+++ src/main/java/dbdr/domain/chart/dto/response/CognitiveManagementResponse.java
@@ -1,0 +1,8 @@
+package dbdr.domain.chart.dto.response;
+
+public record CognitiveManagementResponse(
+        Long id,
+        boolean cognitiveHelp, // 의사소통 도움 여부
+        String cognitiveNote // 인지 관리 특이사항
+) {
+}
